# _potential_nodes_from: invert `as_set` and `patch_requirements`