Test permalinks:
DREAD: Da3N_L4qMzyCzJ0KWIRiEuMOO6C_f_zV
ECHOES: DZQlfQMM47o8ilgMOM9towp4jKE0R9owABzg
PRIME: Deo9W6GlEro8ilgMjaHcfjh85lO6Q9O4AAPW

### Before
DREAD: 23.653s
ECHOES: 5.216s
PRIME: 4.796s

### After
DREAD: 12.953s (82% faster)
ECHOES: 3.467s (50% faster)
PRIME: 2.021s (137% faster)